Added a map for contractors spaces within Battery area below Clay.
Tried generating curb ramp map but all ramps for an intersection have the same lat/lon, so eight objects stack on one another in middle of intersections. The records contain corner and orientation hints, but distance / intersection dimensions are unknown -unless we integrate with sf_streets_polygons.tsv, which would be a big chore. And in the end, I might only discover that nearly every corner is covered, which is what I was seeing.